requirements.txt
fastapi
uvicorn
pydantic
requests
together
Pillow
pyzbar
pytesseract
